requirements.txt
requests
aiohttp
pandas
numpy
python-dotenv
flask
flask-cors
streamlit
plotly
waitress
pyyaml
